+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - Fix crashes when parsing non-UTF8-decodable string literals.
 - \N escape sequence in string literals will no longer cause a syntax error.
+- Module parsing APIs now accept an additional `enable_type_comment` argument, controlling whether to parse type comments or not.
 
 ## 0.1.2 (2021-07-20)